Changed movement

+++ src/utils.c
@@ -6,18 +6,18 @@
 /*   By: ajanse <ajanse@student.42.fr>                +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/22 15:20:51 by ajanse        #+#    #+#                 */
-/*   Updated: 2022/04/22 15:21:20 by ajanse        ########   odam.nl         */
+/*   Updated: 2022/05/03 12:43:41 by ajanse        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <math.h>
 
-float degToRad(int a) 
+float degToRad(float a) 
 { 
 	return a * M_PI / 180.0;
 }
 
-int FixAng(int a)
+float FixAng(float a)
 {
 	if(a > 359)
 		a -= 360;
